dyUnzoom {dygraphs}	R Documentation

The dyUnzoom plugin adds an "Unzoom" button to the graph when it's displaying in a zoomed state (this is a bit more discoverable than the default double- click gesture for unzooming).

Description

The dyUnzoom plugin adds an "Unzoom" button to the graph when it's displaying in a zoomed state (this is a bit more discoverable than the default double- click gesture for unzooming).

Usage

dyUnzoom(dygraph)

Arguments

dygraph
Dygraph to add plugin to

Value

Dygraph with Unzoom plugin enabled

Examples

library(dygraphs)
dygraph(ldeaths) %>%
  dyRangeSelector() %>%
  dyUnzoom()

[Package dygraphs version 1.1.1.6 Index]