dygraphs-exports {dygraphs}	R Documentation

dygraph exported operators and S3 methods

Description

The following functions are imported and then re-exported from the dygraphs package to avoid listing the magrittr and zoo packages as Depends of dygraphs.

[Package dygraphs version 1.1.1.6 Index]